pyproject.toml
[project]
name = "decathlon-usecase"
version = "1.0.0"
description = "forecasting usecase"
authors = [
    {name = "miguel",email = "miguelmayhem92@gmail.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "pandas (<3.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "xgboost (>=3.2.0,<4.0.0)",
    "numpy (>=2.4.3,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
